seperate more bodies
