+++ src/generators/bodies/index.ts
@@ -1,1 +1,5 @@
+export * from "./client-operation.js";
+export * from "./common-route-handler-method.js";
+export * from "./is-authentication.js";
+export * from "./is-parameters.js";
 export * from "./route-handler-method.js";
