+++ src/generators/functions/is-parameters.ts
@@ -1,6 +1,7 @@
 import * as models from "../../models/index.js";
 import { toCamel, toPascal } from "../../utils/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
+import { generateIsParametersFunctionBody } from "../bodies/index.js";
 
 export function* generateIsParametersCode(apiModel: models.Api) {
   yield* generateAllFunctions(apiModel);
@@ -31,56 +32,7 @@ function* generateFunction(
     export function ${functionName}(
       requestParameters: Partial<Record<keyof ${typeName}, unknown>>,
     ): requestParameters is ${typeName} {
-      ${generateFunctionBody(apiModel, operationModel)}
+      ${generateIsParametersFunctionBody(apiModel, operationModel)}
     }
-  `;
-}
-
-function* generateFunctionBody(
-  apiModel: models.Api,
-  operationModel: models.Operation,
-) {
-  const allParameterModels = [
-    ...operationModel.queryParameters,
-    ...operationModel.headerParameters,
-    ...operationModel.pathParameters,
-    ...operationModel.cookieParameters,
-  ];
-
-  for (const parameterModel of allParameterModels) {
-    const parameterSchemaId = parameterModel.schemaId;
-    const parameterTypeName =
-      parameterSchemaId == null
-        ? parameterSchemaId
-        : apiModel.names[parameterSchemaId];
-    if (parameterTypeName == null) {
-      continue;
-    }
-
-    const isFunctionName = `is${parameterTypeName}`;
-
-    const parameterPropertyName = toCamel(parameterModel.name);
-
-    if (parameterModel.required) {
-      yield itt`
-        if(requestParameters.${parameterPropertyName} === undefined) {
-          return false;
-        }
-      `;
-    }
-
-    yield itt`
-      if(
-        !${isFunctionName}(
-          requestParameters.${parameterPropertyName}
-        ) === undefined
-      ) {
-        return false;
-      }
-    `;
-  }
-
-  yield itt`
-    return true;
   `;
 }
